Perform global-local index conversion on client
This leads to a network traffic reduction
as now only local integer indices have to be
transmitted instead of global long indices.

+++ src/it/scala/glint/spark/GranularBigWord2VecMatrixSpec.scala
@@ -40,7 +40,7 @@ class GranularBigWord2VecMatrixSpec extends FlatSpec with SparkTest with Matcher
     try {
       val granularModel = new GranularBigWord2VecMatrix(model, 10000)
       val random = new Random(1)
-      val rows = Array.fill(10000)(Array.fill(random.nextInt(15))(random.nextInt(vocabCns.length).toLong))
+      val rows = Array.fill(10000)(Array.fill(random.nextInt(15))(random.nextInt(vocabCns.length)))
 
       val result = whenReady(granularModel.pullAverage(rows)) {
         identity
