+++ src/main/scala/glint/messages/server/request/PullAverageRows.scala
@@ -5,4 +5,4 @@ package glint.messages.server.request
   *
   * @param rows The array of row indices to average
   */
-private[glint] case class PullAverageRows(rows: Array[Array[Long]]) extends Request
+private[glint] case class PullAverageRows(rows: Array[Array[Int]]) extends Request
