+++ src/main/scala/glint/messages/server/request/PullMultiply.scala
@@ -7,4 +7,4 @@ package glint.messages.server.request
   * @param startRow The start row index of the matrix, to support multiplication with only a part of the partial matrix
   * @param endRow The exclusive end row index of the matrix
   */
-private[glint] case class PullMultiply(vector: Array[Float], startRow: Long, endRow: Long) extends Request
+private[glint] case class PullMultiply(vector: Array[Float], startRow: Int, endRow: Int) extends Request
