+++ src/main/scala/glint/messages/server/request/PullNormDots.scala
@@ -6,4 +6,4 @@ package glint.messages.server.request
   * @param startRow The start index of the range of rows whose own partial dot products to get
   * @param endRow The exclusive end index of the range of rows whose own partial dot products to get
   */
-private[glint] case class PullNormDots(startRow: Long, endRow: Long) extends Request
+private[glint] case class PullNormDots(startRow: Int, endRow: Int) extends Request
